refactor(ai): reorganize training/ to ai/, rl/ to train/
- training/ → ai/: reflects full AI pipeline (sim, train, bot)
- rl/ → train/: directory contains SFT + RL, not just RL
- Updated all imports: rl.network → train.network, etc.
- Updated docstrings with new module paths
- No TS-side changes needed (no references to training/)
